Renamed 'rebase.js' to 'rebase_util'
Need to use the name 'rebase' for a Rebase value type
